+++ package/todo/lib/model/todo.dart
@@ -1,0 +1,22 @@
+class ToDo {
+  String? id;
+  String? todoText;
+  bool isDone;
+
+  ToDo({
+    required this.id,
+    required this.todoText,
+    this.isDone = false,
+  });
+
+  static List<ToDo> todoList() {
+    return [
+      ToDo(id: '01', todoText: 'Morning Excercise', isDone: true),
+      ToDo(id: '02', todoText: 'Buy Groceries', isDone: true),
+      ToDo(id: '03', todoText: 'Check Emails'),
+      ToDo(id: '04', todoText: 'Check Emails'),
+      ToDo(id: '05', todoText: 'Check Emails'),
+      ToDo(id: '06', todoText: 'Check Emails'),
+    ];
+  }
+}
